service-planner-agent/backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
python-dotenv
sse-starlette
